+++ config/test.exs
@@ -2,8 +2,8 @@ import Config
 
 config :elixir_lokalise_api,
   api_token: "fake_token",
-  oauth2_client_id: {:system, "OAUTH2_CLIENT_ID"},
-  oauth2_client_secret: {:system, "OAUTH2_CLIENT_SECRET"},
+  oauth2_client_id: "fakeoauth2clientid",
+  oauth2_client_secret: "fakeoauth2clientsecret",
   http_client: ElixirLokaliseApi.HTTPClientMock,
   request_options: [
     receive_timeout: 5_000,